fix: move getServerFunctionMeta export to @solidjs/start
Refs: #1772, #1365
