+++ packages/start/src/index.tsx
@@ -1,6 +1,6 @@
 // @refresh skip
+export { default as clientOnly } from "./shared/clientOnly";
 export { GET } from "./shared/GET";
 export { HttpHeader } from "./shared/HttpHeader";
 export { HttpStatusCode } from "./shared/HttpStatusCode";
-export { default as clientOnly } from "./shared/clientOnly";
-
+export { getServerFunctionMeta } from "./shared/serverFunction";
