+++ packages/start/src/server/index.tsx
@@ -1,9 +1,11 @@
 // @refresh skip
-export { StartServer } from "./StartServer";
 export { createHandler } from "./handler";
-export { getServerFunctionMeta } from "./serverFunction";
+export { StartServer } from "./StartServer";
 export type {
   APIEvent,
   APIHandler, Asset, ContextMatches, DocumentComponentProps, FetchEvent, HandlerOptions, PageEvent, ResponseStub, ServerFunctionMeta
 } from "./types";
+import { getServerFunctionMeta as getServerFunctionMeta_ } from "../shared/serverFunction";
 
+/** @deprecated */
+export const getServerFunctionMeta = getServerFunctionMeta_;
