+++ tests/server-function/src/functions/use-server-function-meta.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { getServerFunctionMeta } from "@solidjs/start";
+
+export function serverFnWithMeta() {
+  return typeof getServerFunctionMeta()?.id;
+}
